requirements.txt
altair==4.2.2
appdirs==1.4.4
attrs==22.2.0
blinker==1.5
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cymem==2.0.7
dataclasses-json==0.5.7
decorator==5.1.1
doccano-client==1.2.7
docker-pycreds==0.4.0
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
langcodes==3.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
murmurhash==1.0.9
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.9.1
rich==13.3.1
semver==2.13.0
sentry-sdk==1.16.0
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-streamlit==1.0.5
srsly==2.4.6
streamlit==1.19.0
thinc==8.1.8
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
wandb==0.13.10
wasabi==1.1.1
watchdog==2.3.1
zipp==3.15.0
